Added support for stock genotypes.
This commit adds support for obtaining ASCII stock genotypes
from Chado.  The non ASCII version utilizes the FlyBase
specific format with '@' and IDs which is not very user friendly.

It also switches from using an if then else branch to a case
search for resolving which columns and tables to use for symbol
and fullname queries.