Moved flybase.gene mat view to gene.gene table.
Using a materialized view caused issues when trying to drop the flybase
and gene schemas.  PostgreSQL attempts to refresh the view when dumping
so this requires a refresh during import.  Since we are not
moving the feature table (the source of the view) this causes
problems

This commit moved flybase.gene to gene.gene and establishes foreign keys
between the insertion and allele tables and this new table.