Added reg region flag.  Changed id field handling.
This commit adds the 'contains_regulatory_region' field to the
gene.allele table.  This is used to distinguish between construct
alleles that contain a regulatory region of a gene and those
that contain/affect the coding region of the gene.

The ID field was changed to an ALTER TABLE statement from an
explicit call to nextval().  This was to allow UNION statements
to filter out duplicates.  Including an auto-incrementing value
as a part of the UNION prevents this.